feat(documentation): see below. bumped all to 1.2.0 - fixed kover badges not appearing, referring to new version and name, lowered coverage min bound to 60 (need to increase kafka coverage while enabling custom min bound for sub projects) - kafka: added readme, added kdocs, added license credit to original creators - added pr title action
