+++ gradle/wrapper/gradle-wrapper.properties
@@ -3,4 +3,4 @@ distributionPath=wrapper/dists
 distributionUrl=https\://services.gradle.org/distributions/gradle-8.2.1-bin.zip
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-org.gradle.jvmargs=-Xmx4g -XX:MaxMetaspaceSize=2g
+org.gradle.jvmargs=-Xmx2g -XX:MaxMetaspaceSize=1g
