+++ ktor-client-circuit-breaker/README.md
@@ -1,6 +1,6 @@
 # Circuit Breaker Plugin for Ktor Client
-
-<a href="file:/Users/ido/IdeaProjects/flax-ktor-plugins/ktor-client-circuit-breaker/build/reports/kover/html/index.html">![koverage](https://img.shields.io/badge/94.07-green?logo=kotlin&label=koverage&style=flat)</a>
+![Language](https://img.shields.io/github/languages/top/flaxoos/flax-ktor-plugins?color=blue&logo=kotlin)
+<a href="file:/Users/ido/IdeaProjects/flax-ktor-plugins/ktor-client-circuit-breaker/build/reports/kover/html/index.html">![koverage](https://img.shields.io/badge/93.58-green?logo=kotlin&label=koverage&style=flat)</a>
 
 The Circuit Breaker pattern is a crucial part of modern software architecture, allowing you to gracefully handle potential failures when communicating with external services. The Ktor Circuit Breaker plugin is an implementation of this pattern, providing you with an easy way to incorporate this mechanism into your Ktor clients.
