--- ktor-server-kafka/README.adoc
@@ -1,64 +1,0 @@
-= Ktor Kafka Feature
-
-== Links
-
-* https://gamov.dev/ktor-kafka-git[Self-link]
-* https://gamov.dev/ktor-kafka[a workshop walkthrough]
-
-== Usage
-
-* add dependency and jitpack repo
-+
-[source,groovy]
-.build.gradle
-----
-repositories {
-  maven { url 'https://jitpack.io' }
-}
-
-dependencies {
-  implementation 'com.github.gAmUssA:ktor-kafka:main-SNAPSHOT'
-}
-----
-+
-
-[source,kotlin]
-.build.gradle.kts
-----
-repositories {
-    maven {
-        url = uri("https://jitpack.io")
-    }
-}
-
-dependencies{
-    implementation("com.github.gAmUssA:ktor-kafka:main-SNAPSHOT")
-}
-----
-
-* in `Application.module`
-+
-[source,kotlin]
-.Application.kt
-----
-import io.confluent.developer.ktor.Kafka
-import io.confluent.developer.ktor.newTopic
-
-fun Application.module(){
-
-    //install Kafka feature
-    install(Kafka) {
-        configurationPath = "src/main/resources/kafka.conf"
-        topics = listOf(
-            newTopic("myTopic") {
-                partitions = 3
-                replicas = 1
-            }
-        )
-    }
-}
-----
-
-
-
-
