+++ ktor-server-kafka/build.gradle.kts
@@ -1,9 +1,9 @@
+
+
 plugins {
     id("ktor-server-plugin-conventions")
     alias(libs.plugins.kotlin.serialization)
 }
-
-version = "0.0.1"
 
 repositories {
     mavenCentral()
@@ -14,17 +14,22 @@ repositories {
         url = uri("https://packages.confluent.io/maven/")
     }
 }
+
+tasks.matching { it.name.contains("native", ignoreCase = true) }.configureEach {
+    enabled = false
+}
+
 kotlin {
     sourceSets {
         jvmMain {
             dependencies {
                 api(libs.kafka.clients)
                 api(libs.kafka.streams)
-                api(libs.kafka.schema.registry.client)
-                api(libs.kafka.avro.serializer)
-                api(libs.avro4k.core)
-                implementation(libs.arrow.core)
-                implementation(libs.arrow.fx.coroutines)
+                implementation(libs.kafka.schema.registry.client)
+                implementation(libs.kafka.avro.serializer)
+                implementation(libs.avro4k.core)
+                implementation(libs.ktor.client.contentNegotiation)
+                implementation(libs.ktor.serialization.kotlinx.json)
             }
         }
         jvmTest {
