+++ ktor-server-kafka/src/jvmMain/kotlin/io/github/flaxoos/ktor/server/plugins/kafka/components/Producer.kt
@@ -1,0 +1,8 @@
+package io.github.flaxoos.ktor.server.plugins.kafka.components
+
+import io.github.flaxoos.ktor.server.plugins.kafka.KafkaRecordKey
+import io.github.flaxoos.ktor.server.plugins.kafka.Producer
+import org.apache.avro.generic.GenericRecord
+import org.apache.kafka.clients.producer.KafkaProducer
+
+internal fun Map<String, Any?>.createProducer(): Producer = KafkaProducer<KafkaRecordKey, GenericRecord>(this)
