+++ ktor-server-rate-limiting/build.gradle.kts
@@ -2,8 +2,6 @@ plugins {
     id("ktor-server-plugin-conventions")
 }
 
-version = "1.1.0"
-
 kotlin {
     sourceSets {
         commonMain {
