+++ ktor-server-rate-limiting/src/commonTest/resources/logback-test.xml
@@ -8,5 +8,5 @@
         <appender-ref ref="STDOUT"/>
     </root>
     <logger name="io.netty" level="INFO"/>
-    <logger name="io.flax" level="DEBUG"/>
+    <logger name="io.github.flaxoos" level="DEBUG"/>
 </configuration>
